Rename: editUserValidation, essentialValidation 통합
useValidation
