+++ webapp/src/service/user/user.validation.js
@@ -1,4 +1,21 @@
-const essentialValidation = ({
+export const editUserValidation = ({ nickname, techSkills }) => {
+  const validateErrors = {
+    nickname: '',
+    techSkills: '',
+  };
+
+  if (!nickname) {
+    validateErrors.nickname = '닉네임이 입력되지 않았습니다. ';
+  }
+
+  if (techSkills.length === 0) {
+    validateErrors.techSkills = '최소 한 개 이상의 언어를 선택해주세요.';
+  }
+
+  return validateErrors;
+};
+
+export const essentialValidation = ({
   nickname,
   profileImage,
   techSkills,
@@ -35,5 +52,3 @@ const essentialValidation = ({
 
   return validateErrors;
 };
-
-export default essentialValidation;
